waveshare_eink/requirements.txt
Pillow
paho-mqtt
spidev
RPi.GPIO
